+++ CHANGELOG.md
@@ -9,8 +9,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 - Configuration to exclude custom doktypes when automatically generating the breadcrumb (#84)
-- Content Object (cObject) :ts:`SCHEMA` to add Types to Schema via TypoScript (#88)
-  Thanks to `Daniel Siepmann <https://daniel-siepmann.de/about-me.html>`_
+- Content Object (cObject) `SCHEMA` to add types via TypoScript (#88)
+  Thanks to [Daniel Siepmann](<https://daniel-siepmann.de/about-me.html>)
 
 ## [2.2.2] - 2022-01-02